Assignments/Assignment 04/pyproject.toml
[project]
name = "llm-evaluation"
version = "0.1.0"
description = "Evaluation scripts for benchmarking LLMs."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.70.0",
    "matplotlib>=3.10.8",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[project.optional-dependencies]
dev = [
    "black>=26.3.1",
    "mypy>=1.20.0",
    "ruff>=0.15.9",
]

[tool.ruff]
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
